# JSON optimizations, cleanups and listsendpays optimizations