… common/test/run-json.c
… common/test/run-json_remove.c
… lightningd/test/run-jsonrpc.c
… plugins/pay.c
… wallet/db.c