@fiatjaf reported in #3941 that `listspays` was slow, and indeed, it was.

I created 50,000 payments, and ran some tests.

* With no optimization:
    * `listsendpays` takes 0.983 seconds
    * `listpays` takes 52.415 seconds
* With `-O3 -flto`:
    * `listsendpays` takes 0.628 seconds
    * `listpays` takes 43.104 seconds

After these optimizations (mostly FIXMEs!) the results are:

* With no optimization:
    * `listsendpays` takes 0.676 seconds
    * `listpays` takes 1.545 seconds.
* With `-O3 -flto`:
    * `listsendpays` takes 0.416 seconds
    * `listpays` takes 0.971 seconds.