@rustyrussell 2020-08-20
... and the other test where I made the same damn mistake...